Version 4.17

# -----BEGIN PGP SIGNATURE-----
#
# iQIzBAABCAAdFiEEh+Zzr4P2w6DDRMjD9KoinU1YwkUFAmQ9MH4ACgkQ9KoinU1Y
# wkU/SQ/9HBM/1rrvx3gbwpCjD3b1Yh15NgQ+uyzJUHSfnVU3k4u8XARB4enZvzJg
# 2Qtyss1VZPdH1cTqdCil8l1ymMeYUUfI0NQnQKgo4aiwJ7uI5HPVMj5hhHzJdAbr
# u8PzyI/oSUJK5M1EZZVvgUNlgEVm9ZVzX9hzzxXkPFLTDhlwmJFh1fLt0p3M7i6f
# hkWhIdn+ywHpBeWoQbw2uB3nVJrzBNh+IiK+6DGBQa/aq89TTs3Igv6HgMo4/gAL
# JOMWOeJ3AIFIbMXcoCi/oK8cCYw+AO1gRxBsTB1IXzGck6eDyNYuBY/vJbSZmou6
# eDgJUqBK2UDEMMa9zoeoup7OrXJ9QNPd1f4Zvj6oZ3xdlICUxNOKDz+P0V8u/yvH
# DZT5Aqt/xyrmrWVwC4BInKzjFsb8CPM72Isdl+Bz1eXBndOdYBEFftbFfcnrfu6e
# vbSjmjlH+vXZEzivK2z1t+KgJnhkJW8JjJrnHRD1c3hAufTAwYavpjjcCmToXYsh
# hqdJBsVqCYg6vT1hki0NFVkTv1YGdBYXCRTFn4Qhs+agMGmX9pihf5blspaxx8tw
# d9Sw+Rylu1oPagAJTBgHQaoZcjwDydP7bPoeFOXcUh9oop5vSrT4BgxFEUWcVQlZ
# JIizV0gxyoEv6GrBjD0y1AkbLE1XcoAuDiPlS+5YvLQX5Nfid68=
# =khDm
# -----END PGP SIGNATURE-----
# gpg: Signature made 17.04.2023 13:41:50 rodkowoeuropejski czas letni
# gpg:                using RSA key 87E673AF83F6C3A0C344C8C3F4AA229D4D58C245
# gpg: Can't check signature: No public key

+++ weblate/trans/tasks.py
@@ -20,6 +20,7 @@ from django.utils.translation import ngettext, override
 from weblate.addons.models import Addon
 from weblate.auth.models import User, get_anonymous
 from weblate.lang.models import Language
+from weblate.machinery.base import MachineTranslationError
 from weblate.trans.autotranslate import AutoTranslate
 from weblate.trans.exceptions import FileParseError
 from weblate.trans.models import (
@@ -149,36 +150,40 @@ def commit_pending(hours=None, pks=None, logger=None):
 
 
 @app.task(trail=False)
-def cleanup_project(pk):
+def cleanup_component(pk):
     """
-    Perform cleanup of project models.
+    Perform cleanup of component models.
 
     - Remove stale source Unit objects.
     - Update variants.
     """
     try:
-        project = Project.objects.get(pk=pk)
-    except Project.DoesNotExist:
+        component = Component.objects.get(pk=pk)
+    except Component.DoesNotExist:
         return
 
-    for component in project.component_set.filter(template="").iterator():
-        # Remove stale variants
-        with transaction.atomic():
-            component.update_variants()
+    # Skip monolingual components, these handle cleanups based on the template
+    if component.template:
+        return
 
-        translation = component.source_translation
-        # Skip translations with a filename (eg. when POT file is present)
-        if translation.filename:
-            continue
-        with transaction.atomic():
-            # Remove all units where there is just one referenced unit (self)
-            deleted, details = (
-                translation.unit_set.annotate(Count("unit"))
-                .filter(unit__count__lte=1)
-                .delete()
-            )
-            if deleted:
-                translation.log_info("removed leaf units: %s", details)
+    # Remove stale variants
+    with transaction.atomic():
+        component.update_variants()
+
+    translation = component.source_translation
+    # Skip translations with a filename (eg. when POT file is present)
+    if translation.filename:
+        return
+
+    # Remove all units where there is just one referenced unit (self)
+    with transaction.atomic():
+        deleted, details = (
+            translation.unit_set.annotate(Count("unit"))
+            .filter(unit__count__lte=1)
+            .delete()
+        )
+        if deleted:
+            translation.log_info("removed leaf units: %s", details)
 
 
 @app.task(trail=False)
@@ -390,10 +395,18 @@ def auto_translate(
         auto = AutoTranslate(
             user, translation, filter_type, mode, component_wide=component_wide
         )
-        if auto_source == "mt":
-            auto.process_mt(engines, threshold)
-        else:
-            auto.process_others(component)
+        try:
+            if auto_source == "mt":
+                auto.process_mt(engines, threshold)
+            else:
+                auto.process_others(component)
+        except MachineTranslationError as error:
+            translation.log_error("failed automatic translation: %s", error)
+            return {
+                "translation": translation_id,
+                "message": _("Automatic translation failed: %s") % error,
+            }
+
         translation.log_info("completed automatic translation")
 
         if auto.updated == 0:
